backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1
tortoise-orm==0.20.1
asyncpg==0.29.0
aiomqtt==2.3.0
orjson==3.10.7
httpx==0.27.2
cryptography==43.0.1
python-multipart

